requirements.txt
addict==2.4.0
autocommand==2.2.2
backports.tarfile==1.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
groundingdino-py==0.4.0
huggingface-hub==0.29.2
humanfriendly==10.0
idna==3.4
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.4
kiwisolver==1.4.7
lit==15.0.7
MarkupSafe==2.1.5
matplotlib==3.9.4
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
onnx==1.17.0
onnxruntime==1.19.2
onnxruntime-gpu==1.18.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
protobuf==6.30.0
pycocotools==2.0.7
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.1
safetensors==0.5.3
scipy==1.13.1
six==1.17.0
supervision==0.6.0
sympy==1.13.1
terminaltables==3.1.10
timm==0.4.12
tokenizers==0.13.3
tomli==2.2.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.67.1
transformers==4.30.2
triton==2.0.0
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==1.26.13
yapf==0.40.2
zipp==3.21.0
